+++ JuceLibraryCode/modules/juce_core/streams/juce_OutputStream.cpp
@@ -252,12 +252,12 @@ bool OutputStream::writeText (const String& text, const bool asUTF16,
     return true;
 }
 
-int64 OutputStream::writeFromInputStream (InputStream& source, int64 numBytesToWrite)
+int OutputStream::writeFromInputStream (InputStream& source, int64 numBytesToWrite)
 {
     if (numBytesToWrite < 0)
         numBytesToWrite = std::numeric_limits<int64>::max();
 
-    int64 numWritten = 0;
+    int numWritten = 0;
 
     while (numBytesToWrite > 0)
     {
